WeMediaPublishNewsCommit

+++ heima-leadnews-service/heima-leadnews-wemedia/src/main/java/com/heima/wemedia/interceptor/WmInterceptor.java
@@ -30,7 +30,7 @@ public class WmInterceptor implements HandlerInterceptor {
         // 存入threadLocal中
         if (userId != null && !userId.isEmpty()) {
             log.info("当前用户：{}", userId);
-            WmThreadLocalUtil.setCurrentId(Long.valueOf(userId));
+            WmThreadLocalUtil.setCurrentId(Integer.valueOf(userId));
         }
         // 放行
         return true;
