+++ heima-leadnews-utils/src/main/java/com/heima/utils/common/WmThreadLocalUtil.java
@@ -7,13 +7,13 @@ package com.heima.utils.common;
  */
 public class WmThreadLocalUtil {
 
-    public static ThreadLocal<Long> threadLocal = new ThreadLocal<>();
+    public static ThreadLocal<Integer> threadLocal = new ThreadLocal<>();
 
-    public static void setCurrentId(Long id) {
+    public static void setCurrentId(Integer id) {
         threadLocal.set(id);
     }
 
-    public static Long getCurrentId() {
+    public static Integer getCurrentId() {
         return threadLocal.get();
     }
